place onLoadMore to separate file
