+++ src/modules/createBtn.ts
@@ -11,9 +11,7 @@ export function createBtn() {
     wrapper.appendChild(returnBtn);
     returnBtn.addEventListener('click', () => {
         wrapper.innerHTML = '';
-        console.log(`The old screen was ${getStateValue('screen')}`);
         setStateValue('screen', 'albums');
-        console.log(`The new screen is ${getStateValue('screen')}`);
     });
     return returnBtn;
 }
