+++ src/renderGallery.ts
@@ -24,21 +24,15 @@ export async function renderGallery() {
 
     if (Object.keys(photosFromState).length !== 0) {
         photos = photosFromState;
-        console.log('The very first Load');
     } else {
         photos = {
             [albumId]: await getData(
                 `https://jsonplaceholder.typicode.com/albums/${albumId}/photos`
             ),
         };
-        console.log('Second Load');
     }
 
     try {
-        console.log(`ID STATE VALUE: ${idFromState}`);
-        console.log(
-            `ID FROM LOCALSTORAGE: ${Number(localStorage.getItem('albumID'))}`
-        );
         wrapper.innerHTML = '';
         const galleryWrapper = createElement({
             tag: 'div',
@@ -46,9 +40,6 @@ export async function renderGallery() {
         });
         wrapper.appendChild(galleryWrapper);
         const restoredAlbums = JSON.parse(localStorage.getItem('albumsInfo'));
-        console.log(restoredAlbums);
-        console.log(albumId);
-        console.log(restoredAlbums[albumId].title);
         const galleryInfo = createElement({
             tag: 'div',
             className: 'gallery__info',
